tast/security.ToolchainOptions: whitelist vdso .so files
Add directories containing linux-gate.so files to whitelist.

The ToolchainOptions test was failing on Caroline with kernel 4.19 due
to linux-gate.so files (in /lib/modules/*/vdso/vdso*.so.dbg/) not being built with
RELRO and non-executable stack.

The linux-gate.so file is just a stripped version of vdso32/64.so which
is built together with the Linux kernel without the -z,relro flag, so it
is out of our control.

BUG=b:145750675
TEST=tast run $DUT_IP security.ToolchainOptions

Change-Id: I5a5cc698f9802cd6e95ce8d15d11aea1957fa0b6
Signed-off-by: Michal Stanek <mst@semihalf.corp-partner.google.com>