Remove metrics.SetConsent
metrics.SetConsent was the old way of getting metrics consent (mostly so
that we could collect crashes). It wasn't very reliable.

Replace the last use with SetUpCrashTest's consent-getter. This is a bit
long-winded, because it requires us to log in and then log back out, but
it should be more reliable, and mean we don't have to maintain 2 consent
systems.

It also fixes problems where the ExtraArgs being passed to Chrome were
being ignored. (https://crbug.com/1042483)

BUG=chromium:1042483
Fixed: chromium:1042483
TEST=tast -verbose run --extrauseflags chrome_internal  my_crbook ui.ChromeCrashNotLoggedIn.*

Change-Id: I62029e78ab037c5fece8feeecf6d73d523c310cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2004088
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>