Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > (4 comments)
> > 
> > + Kirtika, FYI: I think I realized why you need these hacks:
> > 
> > https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/refs/heads/master/net-wireless/iwlwifi_rescan/files/pci-rescan-to-revive-wifi.sh#145
> > (restart wpasupplicant / restart shill)
> > 
> > It's probably because of https://crbug.com/1029162
> 
> Yikes, and thanks for the link.

Wait, reading closer, I am not following. Your bug description in crbug.com/1029162 was specific to wpa_supplicant-2.8. But I had to put the hack in there, during ~2017/2018 when we were at wpa_supplicant-2.5 (though not calling it that). What am I missing? 

I agree with your comment in bug description that we need an integration test for this.

Patch-set: 3