Create Core of Tast UI Automation Library
Adds a wrapper for the chrome.automation api in order to make
developping UI tests for ChromeOS easier. The library is defined in
go/tast-ui-auto. It enables searching for and interacting with
chrome.automation AutomationNodes from within Tast. Currently, it is
possible to find nodes, get attributes of nodes, wait for a node to be
added/removed, and left/right click on nodes.

BUG=chromium:982451
TEST=Locally tested on Eve

Change-Id: I0b14040bd63a7e8b4d4d1ff01cd34f495e147cb6