Update patch set 2
Patch Set 2: Verified+1 Auto-Submit+1

> Patch Set 1:
> 
> Routines will be disabled soon, because we are moving to cros_healthd.
> 
> VM tests are not stable enough yet.

I think the VM tests are stable enough to be enabled. If you look at the stainless board commented in the first patchset, the only VM test failure for wilco.VMStart was due to a CL that broke the VM. If this test was enabled, it would have prevented that CL from going through. (http://b/146811529). This is not a flake, but the test doing it's job. There were many runs before this breaking change passed. Therefore, I think the tests that I have in this CL should be enabled as critical.

Patch-set: 2
Reviewer: Gerrit User 1304442 <1304442@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1
Label: Auto-Submit=+1