tast-tests: fix arc.SetBounds to start activities separately
In arc.SetBounds, the test launches one activity, and clicking a button
on the the app invokes to start another activity.

This CL changes to use testing.State.Run to have multiple subtests and
start each activity only in the subtest that uses it.

This change fixes current tast failure due to mismatch of activity name
between arc.Activity instance held in test and actually running activity
in ARC container.

BUG=chromium:1042617
TEST=tast run DUT arc.SetBounds
Change-Id: Ia16dd021c3413e4af5ddb3d7ff22c48bb86feba8