Crostini tast: Recover from Chrome crash in subsequent tests.
Currently if Chrome crashes during a Crostini test, all subsequent tests
using the same precondition will fail if they try to communicate with
Chrome.

This change checks if Chrome is still responsive, and if not re-runs the
precondition to restart both chrome and Crostini.

Also changed errors during cleanup to logs to avoid failing a test due
to failed cleanup (which could run either before or after the test in
question).

BUG=chromium:1043352
TEST=Killed chrome at end of one test (via sigsegv) and ensured
subsequent tests failed without the change, and passed after the change.

Change-Id: I674d98d37192eb21a272cabbb9fc4ee4241effbc