tast-test: Fix arc.PlayStore test and correctly enable arcvm
Play Store sometimes shows a different UI. We need to do some extra
checks to figure out which UI the Play Store is showing.

When we search for a package, Play Store may directly open the page for
the package or may open a search page. We now check for the install
button to figure out which happens.

ArcVM was enabled and has a different version of PlayStore. We search
for both the arc container and arcvm search buttons in order to pick the
correct searchbox.

TEST=tast run <eve> arc.PlayStore arc.PlayStore.vm
BUG=chromium:1044407

Depends-On: chromium-review/2013004
Change-Id: Ifceffc58ffe90a8bf1a64a78b6bc1d370115c2ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2015567
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>