Update patch set 1
Patch Set 1:

The important thing here is to check that the tests to be moved are behaving reasonably well before the move, such as not to destabilize the tests in graphics_perbuild. If they are not very clean, then moving to daily or weekly makes sense. So let me check.

Patch-set: 1
Reviewer: Gerrit User 1000221 <1000221@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>