Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> Maybe triple or quadruple the timeout to be a bit generous? Timeouts are there to abort when we are really sure that things went wrong. If right now we complete a bit "over half" then I am worried that just doubling is too tight.

Sure thing, quadrupled, which should be more than enough for most cases. There a very few runs that make it through 250-300/1000 seeks, however.

My concern is that if the timeout is too long and tests that really /shouldn't/ take so long slowly creep up in runtime, we won't notice until the resource usage is already higher than we want it to be, but if that's something we're not too concerned about here that's fine.

Patch-set: 5
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>