Update patch set 3
Patch Set 3:

(4 comments)

> Thanks for doing this! Took me a while to parse through the test, but I think I understand it now. Added some preliminary comments and will take another look tomorrow.

Let me know if there's some part that was hard to parse. It's definitely not good if it takes a readers a long time to figure out what's happening.

(I'll work on this myself anyway, so feel free to ignore this CL until I upload a new version.)

> IMO, probably fine to leave it as is for now. I plan on doing VirtualEthernetPair after the DHCP changes land, so I can just come back to this and abstract it at a later date.

Ack. Anyway, the virtual-ethernet stuff might be worth extracting at least into a separate function for readability.

> You are the first test to be using hostapd in tast, is it reasonable to ask for some introductory notes? 

I'm not sure what sort of thing you're looking for, but I can try to add a few notes in either the commit message or the code comments.

Patch-set: 3