tast: Use cros_config to check if device has SKU number
Use cros_config to check if a device has a SKU number. If it does,
verify that the SKU number is printed. If it does not, verify that "NA"
is printed.

Also, add a new crosconfig support library for tests that use
cros_config.

BUG=None
TEST=tast run --build $DUT_IP platform.CrosHealthdProbeCachedVpd
TEST=tast run --build $DUT_IP hardware.SensorPresence
TEST=fast_build.sh -T

Change-Id: Idf7740b252eb44439136a9e420e20e0fdec5f1a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2023326
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Jesse Schettler <jschettler@chromium.org>