tast-test: arc.WindowManagerCUJ use unspecified activity
Although all clamshell Chromebooks are in landscape orientation, it is
possible to put one in portrait mode. For example, a Potrait-tablet like
Dru can be put in "clamshell" mode when a keyboard is attached (or by
forcing "clamshell-mode" with our Test API).

When that case occurs, if an activity is launched in "lanscape mode",
the activity will be in "letter-box" mode, and when switched to
immersive mode, it will fail.

This CL, launches the activity in "Unspecified" instead of "Landscape"
to cover both cases: clamshell in portrait and clamshell in landscape.

BUG=None
TEST=Tested on Dru in portrait mode and on Eve in landscape mode.

Change-Id: I567051c9e723426f1c02bf85e8eca9cf949dc951
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2023807
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>