tast-test: arc.WindowManagerCUJ use unspecified activity
Although all clamshell Chromebooks are in landscape orientation, it is
possible to put one in portrait mode. For example, a Potrait-tablet like
Dru can be put in "clamshell" mode when a keyboard is attached (or by
forcing "clamshell-mode" with our Test API).

When that case occurs, if an activity is launched in "lanscape mode",
the activity will be in "letter-box" mode, and when switched to
immersive mode, it will fail.

This CL, launches the activity in "Unspecified" instead of "Landscape"
to cover both cases: clamshell in portrait and clamshell in landscape.

BUG=None
TEST=Tested on Dru in portrait mode and on Eve in landscape mode.

Change-Id: I567051c9e723426f1c02bf85e8eca9cf949dc951