[Tast] Add a PolicyBlob to FakeDMS to reduce code in tests.
Since there won't be multiple instances, add a PolicyBlob struct field
to the FakeDMS struct. It is exported, as users may modify it as they
wish. Add convenience functions to quickly update the FakeDMS with
a given set of policies and to return the chrome Options needed to
connect the FakeDMS to Chrome.

BUG=None
TEST=ran existing policy.* tests

Change-Id: I5865a297c3f9c94027d57065710a0cc9cf783ffd