Split ChromeCrash tests between breakpad and crashpad
We are moving from breakpad to crashpad for handling Chrome crashes. An
upcoming experiment will move an increasing # of Chrome sessions to
using crashpad.

Create crashpad and breakpad versions of all tests. This ensures:
1. That crashpad works on all platforms and multiple scenarios before it
   is launched.
2. That tests don't appear flaky because session_manager sometimes
   starts them with crashpad and sometimes with breakpad. The tests will
   now always run with a known crash handler.

Breakpad is known to not work on kukui, so the breakpad tests are
disabled on kukui (which has the "force_crashpad" USE flag).

BUG=chromium:1042993
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook ui.ChromeCrash* on both eve and kukui.

Change-Id: Id4be874e4b59b6973e083bfdfff1c077a5c7e14d