crash: Use the standard crash dir for crash_sender tests.
This is a part of changes to merge sender-related utilities.

To merge sender.SetUp to crash.SetUpCrashTest, it is preferable
not to change the signature of crash.SetUpCrashTest, therefore
we cannot create a temporary directory in the setup function.
Instead we can use the standard crash directories which are
saved by the setup function.

This change rewrites the sender subpackage and existing
crash_sender to use the standard crash directories.

BUG=chromium:1046647
TEST=fast_build.sh -T
TEST=tast run DUT crash.Sender*

Cq-Depend: chromium:2014180
Change-Id: I5f0ebd0094c764588c6bff2037783689f4b756dc