Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> If you look at
> https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2013131
> tests no longer need to handle chrome startup at all.
> 
> This is still useful if we have tests that need to perform a custom chrome setup.
> 
> I am not sure about the Blob on FakeDMS, keeping WritePolicyBlob as is seems more intuitive. ServePolicies can just create a new blob and write it.


Creating a new Policy Blob for every new set of policies makes it hard to use a non-default setup across multiple tests or subtests.  E.g. if we want a precondition with a different policy_user or any of the other dozen options for policy_testserver, each iteration of the test would have to modify the PolicyBlob the same way.

I'm willing to be convinced away from adding Blob to FakeDMS - the main draw is the Chrome login Options, and that could be done by passing in a PolicyBlob - but having functions on FakeDMS/PolicyBlob for clearing/serving policies seems like a good idea to me.

On the other side of the fence, there's also the option of not exporting PolicyBlob and putting methods to modify its various pieces on FakeDMS - then there's no way to get into a state where the PolicyBlob gets modified but not written to the FakeDMS (which this CL's proposal is bad about).

Patch-set: 1