Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > We are already low on tests that can exercise TPM/GSC initialization during OOBE. And soon this will not be the only test that clears the owner. Disabling it now will only sweeps the problem under the rug until it becomes bigger. I'd suggest we create a way to isolate such tests now, or add measures to preserve logs when running such tests (copy logs to outside of encrypted stateful, look for them there).
> 
> I don't think that's the right approach. If nya@ is correct, this test is breaking our ability to diagnose other tests (see https://bugs.chromium.org/p/chromium/issues/detail?id=1045624 for example), which means that its presence negatively impacts our testing coverage for all teams much more than its absence. So I think disabling it is the right approach. That gives time to properly isolate it, then re-enable it without breaking other tests, then add more tests that clear the owner.

If the problem can be repro'd, the tests can still be run individually and all logs collected. I really understand the inconvenience (I deal with inability to retrieve pre-owner-clear logs during debugging maybe more than anybody else, sigh), but disabling an otherwise healthy and useful test in the under-tested area is not the way to go imho. If we disable it now, there will be no incentive to come up with an isolation approach, so we'll just keep it disabled. Let's have a separate suite, or run those tests first, or preserve the logs in the disk areas that survive owner clear.

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>