Update patch set 1
Patch Set 1:

All the tests that reset the TPM to the unowned state will clear the logs if the logs are stored in encrypted stateful. We do have soft-clear-TPM tests, which won't do that, but normal owner clear results in encstateful loss.

So, the solution is to add a new group ans et it for this test? Yes, we can do that instead of this CL.

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>