tast-tests: Avoid policy in ChallengeResponseMount
Fix the flakiness in the ChallengeResponseMount test caused by the
"EphemeralUsersEnabled" device policy being set unexpectedly
(presumably, due to some other test setting this policy and not properly
cleaning the state up).

This CL changes the test to explicitly configure device policy with the
"EphemeralUsersEnabled" being set to false in the beginning of the test.

BUG=b:148005583
TEST=run the test; invert the policy value to "true" and cehck that the test fails
Change-Id: I1612967dc2592c4881d5040f39356be3d9e15307