tast-test: Add check for correct package Arc.PlayStore
Arc.PlayStore recently broke due to a change in the Play Store UI that
caused it to install the wrong package. The error was caught manually
instead of by this tests. By checking the correct package at the end of
the test, we ensure that we won't install the wrong package when the UI
changes again.

BUG=b:141885513
TEST=tast run <eve> arc.PlayStore

Change-Id: Ic3581e788fe1e6ef36b3bc43db1e224368e7b1c8