tast-test: Add check for correct package Arc.PlayStore
Arc.PlayStore recently broke due to a change in the Play Store UI that
caused it to install the wrong package. The error was caught manually
instead of by this tests. By checking for the correct package at the end
of the test, we ensure that the test will actually fail when the UI
changes again.

As part of this change, I made arc.installedPackages public and updated
locations in the code that used "pm list packages" to use this code
instead.

BUG=chromium:1044407
TEST=tast run <eve> arc.PlayStore

Change-Id: Ic3581e788fe1e6ef36b3bc43db1e224368e7b1c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2031478
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>