tast/video: add Play cases and logic for memory comparison
This CL adds three test cases and logic for comparing graphics
framebuffer objects before-after playing a video, failing if
those two are not the same (leak!). For this, a new file and
method CompareGraphicsMemoryBeforeAfter() is added, that
compares the values of 915_gem_framebuffer before and after.

Those values might fluctuate; readStableI915ObjectCount() and
waitForStableReadings() emulate the CPU.go wait-for-idle logic
(but they are _not_ the same: CPU goes to zero, here a moving
average is used).

TEST=tast run ${IP} video.Play.*memcheck on kohaku and atlas
BUG=chromium:1030129

Change-Id: I0307bbd18662372149372bb220c7d4edf1e069b3