tast/video: add Play cases and logic for memory comparison
This CL adds three test cases and logic for comparing the amount
of framebuffer objects of specific dimensions before-after playing
a video, failing if those two are not the same, which would mean
there's a leak in Chrome-minigbm allocated resources.

A new file dri.go is added to interact with the DRI debugfs. A
graphics.Backend interface is added to define what is needed from
that debugfs, with a single I915Backend implementation. This is
used for three methods: CompareGraphicsMemoryBeforeAfter(),
readStableObjectCount() and waitForStableReadings(), that
basically wait for stability of the readings and compare them.

The latter two methods emulate the CPU.go wait-for-idle logic
but they are _not_ the same: CPU goes to zero, here a moving
average is used.

BUG=chromium:1047514
TEST=tast run ${IP} video.Play.*memcheck on kohaku and atlas

Change-Id: I0307bbd18662372149372bb220c7d4edf1e069b3