Tast: ReportDiskUsage: support chromeperf filename
We run ReportDiskUsage in SimpleChrome builders on the chrome waterfall.

These builders need to upload perf results to chromeperf, which expects
a different filename (and eventually format) than the crosbolot name
and format used in Chrome OS builders.

This CL adds support for a custom filename to perf.go as an intermediate
step (crosperf currently supports the same format, but it will be
deprecated).

BUG=chromium:1047454
TEST=fast_build.sh -t chromiumos/tast/local/perf

Change-Id: I0320a14c7b0e32908b8f85f775a149ea439076a4