wificell: hostapd: Add hostapd package for controlling hostapd.
Basic funcionality for running hostapd is ported from Autotest. This
is part of basic network_WiFi_SimpleConnect test port. Not yet all
supported configurations are ported. They'll be added along with more
SimpleConnect testcases. (only aiming for 802.11a/b/g testcases now)
Also, functions of hostapd_cli are also skipped for now but it should
not be hard to add when needed as we do setup ctrl_interface.

BUG=chromium:1034875
TEST=None

Change-Id: I25f046aba98defb16a21141ffe5b7c9419510650