wifi: Add Router object for controling router
wifi.Router is a port of basic SiteLinuxRouter functionalities. It
controls the resources (IP ranges, WiFi interfaces) and spawns hostapd
and dhcpd. This is part of basic network_WiFi_SimpleConnect test port.

Some functionalities are skipped now, and will be ported in later cls
or when used by other WiFi tests:
- avahi setup
- bridging multiple hostapd instance
- collect more system logs on router
- control of monitor interface
- inline TODOs for router setup

BUG=chromium:1034875
TEST=None

Change-Id: I267e83c5b762a0aeda509e3e74b9b7db7621f0fe