network: IwlwifiPCIRescan: wait for shill restart
In current pci_rescan script, we'll restart shill to mitigate
crbug.com/1029162. We should wait for shill to restart before any
further dbus operation or else shill can restart anytime and causing
errors. Also, if the interface is recovered before shill restart, this
test could just end and leaving the restart to harm later tests.

Once the crbug.com/1029162 is fixed and shill restart is removed, the
test will start failing and we need to revert this commit.

BUG=chromium:953497, chromium:1029162, chromium:1048075
TEST=`tast run $DUT network.IwlwifiPCIRescan` on cyan DUT

Change-Id: Ifa24cdceb72e9e8f194b1971ff2f5c97aa452ef5