arc: Add data collector server test
This test is intended to be run during the Android PFQ in order to
collect set of data. At this moment ureadahead packs are generated and
collected for the initial and provisioned boots. Uploaded results will
be used by project-cheets-private/scripts/board_specific_setup.sh
during the finalizing of the generated system image.

We have following coverage in Android PFQ:
  x86 - caroline-pi-android-pfq, coral-pi-android-pfq,
        grunt-pi-android-pfq, reef-pi-android-pfq
  x86_64 - eve-pi-android-pfq
  arm - kevin-pi-android-pfq
  arm64 - not covered, kevin-arc64 is only one used
          once it is promoted to production, Android PFQ
          should have this arch as well.

At this moment device specific calls are ignored (for example camera,
video DSO), considering the impact is relatively small.

TEST=tast run <Container|ARCVM> arc.DataCollector
BUG=b:148557896

Change-Id: I39a65783bc2e0d69003a3b9aad72600a5306a4e2