tast-tests: Verify crash_reporter setup after init.
This is a replacement of the first part of logging_UserCrash autotest.
reporter_startup in logggin_UserCrash is invoked without re-initializing
crash_reporter to catch problems with the default crash reporting setup.
https://chromium.googlesource.com/chromiumos/third_party/autotest/+/0.12.433.B/client/site_tests/logging_UserCrash/logging_UserCrash.py#676

Some test setup such as stashing crash files and setting the
test-in-progress flag file is omitted, because those will not affect
this specific test checkpoints.

BUG=chromium:1047059
TEST=tast run $DUT crash.ReporterStartup

Change-Id: If7356a862115188de28076fd708d88fc0231e186
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2066402
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>