add performance tests for the hotseat
Add a test which drags the hotseat, and another test
which records all hotseat animations.

Hotseat animations should be recorded in all major transitions
to/from:
 - in-app shelf
 - Home Launcher
 - Overview

Test with a constant wait time (2 seconds) and when CPU is idle.

Ideally this will catch a case where the Sys UI itself caused the
reduction in performance or the device is performing poorly in some
other area.

BUG=chromium:1044289
TEST=manual

Change-Id: Ie6bff2b2f4f1e8fe1641d77bc1d7d6f2f967d3eb