arc: Don't use oldCrashes in AppCrash.
The oldCrashes dir is somewhat obsolete, and using it here was causing a
race: If the app crashed before we got the old crashes, we'd mistakenly
count some of the crash files from _this_ run as "old".

Additionally, the SetUp and TearDown methods take care of moving
existing files out of the crash directory, so there's no need to list
the old ones anymore.

BUG=chromium:1039512
TEST=tast --extrauseflags chrome_internal $DUT arc.AppCrash

Change-Id: I538dd56090e531adef989f52a6481e2efdbc2b78