Update patch set 1
Patch Set 1:

The fix CL looks like not yet landed, and is not the revert of the regression.
So, in order to guarantee that the fix works stably, could you wait for several runs after the fix is landed?

Patch-set: 1
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>