Crostini Tast: Split all artifact tests to stable/unstable
Currently only the Sanity test has a stable/unstable split.  Make this
same split to every other test that has an Artifact or derivative
precondition.

Once these run for a while the majority of the stable artifact tests
should be abler to be promoted to the CQ.

BUG=None
TEST=Ran all crostini tests and verified that the unstable tests were
skipped on eve (a stable board)

Change-Id: If63de23c6f5baa525cb314cbd0050beecfb5db4b