Update patch set 12
Patch Set 12:

(2 comments)

Thanks for all your comments. PTAL

Implementation has changed quite a bit since initial review.
After changing density, test takes a screenshot and confirms the size of the square that is drawn on the screen.
The apk which contains a 100x100 black square is here ag/10214246.

Patch-set: 12