Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> (1 comment)
> 
> Thanks for the library.
> 
> In Tast we prefer to avoid calling binaries, if that same behavior can be achieved with < 100 LoC.
> For this particular case, it should be more or less easy to replace the calls to "backlight_tool" with read/write/stat system calls. e.g:
> 
> read /sys/class/backlight/[device name]/brightness
> write /sys/class/backlight/[device name]/brightness
> 
> 
> Examples of Tast code avoid calling external binaries:
> 
> input, which instead of calling "evtest" it calls syscalls: https://source.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/input/
> power RAPL: https://source.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/power/rapl.go

I thought this would be easy, but it looks like converting lux to a brightness might be hard. backlight_tool is more complicated than I imagined...

It seems to be parsing[1] this preference value [2] in a very roundabout way with listeners and delegates etc. Actual conversion seems to happen here [3].

It feels like there should be a simple formula buried in all that code, but every layer seems to do a little bit of massaging on the values that pass through, so I'm not sure...

[1] https://cs.corp.google.com/chromeos_public/src/platform2/power_manager/powerd/policy/ambient_light_handler.cc?dr=CSs&g=0&l=68
[2] https://cs.corp.google.com/chromeos_public/src/platform2/power_manager/powerd/policy/internal_backlight_controller.cc?dr=CSs&g=0&l=185
[3] https://cs.corp.google.com/chromeos_public/src/platform2/power_manager/powerd/policy/ambient_light_handler.cc?dr=CSs&g=0&l=170

Patch-set: 7
Reviewer: Gerrit User 1368712 <1368712@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>