tast-test: add precondition for appcompat arc with GAIA
This enables using GAIA login as part of the arc precondition. It
creates 4 preconditions that will be used in every app compat test. This
also updates the current app compat test to use the preconditions. As
part of this, the Play Store information was added to the apps package.

BUG=b:141885513
TEST=tast run <eve-dualimage> arcappcompat.*

Change-Id: I92c4152cdbf90fe5c082943d4185cef535be44d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2042224
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>