tast/video: Add ChromeVideoWithGuestLogin pre, use it for mem check tests
This CL  adds a new ChromeVideoWithGuestLogin precondition, which
is a ChromeVideo with Guest login (as opposed to a real Test user)
and uses it for the video.Play.*mem_check test cases.

Using the Guest login is better because the activity during- and
post-login is much less, which makes it easier to track memory
consumption.

The <video> needs to be muted to be played w/o user gesture (otherwise
we get the infamous “Uncaught (in promise) DOMException: play() failed
because the user didn’t interact with the document first.”). This is
not an issue for any other test cases running play.go, so the change
is safe.

BUG=chromium:1047514
TEST=tast run ${IP} video.Play.* on kohaku and atlas

Change-Id: Ic8e2470086af9f8e49aa41eaccd69d6ae2853c0a