tast-tests: ensure Files App has completed rendering after launch
The Files App related tests are currently flaky on slow devices because
they do not wait for the Files App to be ready before continuing. They
may try to open the Downloads folder before the folder list is shown.
This CL waits for an internal part of the UI to render.

BUG=chromium:1050255
TEST=tast run <eve> ui.FilesApp*

Change-Id: I05396bb96a5def2e05236703bfeb4ad3b547c875