Add EarlyCrash test.
Test that crashes happen early in boot are properly collected and tagged
with the right metadata tag.

NOTE: This test flakes a _lot_ now -- maybe 3 or 4 of every 10 times.
This is entirely due to consent, and adding a 3 second sleep after the
SetUp call returns brings it down to maybe 1 in 20. I'm submitting this
in the hopes that the frequent flakiness will be useful for debugging
consent issues.

BUG=chromium:1039378
TEST=tast run --extrauseflags chrome_internal $DUT crash.EarlyCrash

Cq-Depend: chromium:2044555,chromium:2048869
Change-Id: Ifde0c1bcacaaedf97c49863304cda76e8f6544ee