tast-tests: Fix workspace insets test in tablet device
The workspace insets failure in kukui(krane & kohaku) board. The
reason is that the display run landscape orientation in rotation
mode but the width and height of display is fixed. The test rely
on display info so that it can't get the correct workspace insets.

This CL add display orientation detection before workspace test.
It will rotate the display to 0 angle first.

BUG=b:141978622
TEST=tast -verbose run my_crbook arc.CompanionLibrary

Change-Id: I51431a2066f484c8ee3db7903e8d8d116aae6d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2046726
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Shengsong Tan <sstan@google.com>
Commit-Queue: Shengsong Tan <sstan@google.com>