Update patch set 12
Patch Set 12:

(2 comments)

Thanks for the update. The code looks good, with a few nits.

What I'm not sore sure is whether we should merge this file with rapl.go.

As an example, in rapl.og we have a function called ReportPerfMetrics [1], which is almost identical to func Snapshot().

And from a user point of view, it is not clear when the user should use "chromiumos/tast/local/power" and when "chromiumos/tast/local/power/powercap". Both have similar goals.

I'm not against using "chromiumos/tast/local/power/powercap", but as it is now, it seems that we should merge both of them either in "chromiumos/tast/local/power/powercap" or in "chromiumos/tast/local/power"


[1]: http://cs/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/power/rapl.go

Patch-set: 12
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>