test/video: Add Play and Seek cases for the alternative decoder impl
This CL adds the appropriate video.{Play,Seek}.*exp test cases to exercise
the video decoder implemementation. Currently there's two video decoder
implementations, the "old" and the "new" :-) -- these new entries are
meant to be testing the one not in use by default by the user (the "alt")
for the web cases, that is nonetheless used for other reasons (think
Pepper or ARC).

Currently the only difference from the code perspective is the
precondition: ChromeVideoVD() ISO ChromeVideo().

BUG=chromium:1050248
TEST=tast run ${IP} video.Play.*exp video.Seek.*exp on atlas.

Change-Id: I692db06601425b7e34f0debeba24e0e384ed4f15