platform.CrosConfigFS: use new cros_configfs command instead of cros_config
We are migrating the ChromeOS ConfigFS related functionality to a new
command (cros_configfs) instead of lumping it all into cros_config, to
keep functionalities separated.

This new command also allows for unmount, so we can get rid of all of
the logic here that unmounts the private and v1 directories.

BUG=chromium:1041544
TEST=test passes with required dependencies applied

Cq-Depend: chromium:2042644
Change-Id: I1e6e7d2328292750b7aab1184813d44204d7c18f