tast-test: arc.KillProcess handle error correctly
This CL works-around a possible race in the process library.
When the list of processes is fetched, it might be possible that by the
time we want to query something of that a process, that process no
longer exists.

BUG=chromium:1050856
TEST=tast run $DUT arc.KillProcess (run Ok 5/5)

Change-Id: Ib30cdc139139fbed54c9b9dcb75f4fdbb0f2719f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2050668
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>