tast-test: Add ShillInitLoginGuestScript test
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockFlimflam instead of shill to run
the login guest test. MockFlimflam provides a subset of shill DBus API
so that we can track "dbus-send" invocations sent by the shill init
scripts.  These tests are moved to run on real shill in the new Tast
test, because it's possible to observe the expected behaviors without
the mock and that allows for checking the status of the profile stack as
part of these tests.

BUG=chromium:1026901
TEST=run tast network.ShillInitLoginGuestScript

Change-Id: I2af2350f13db7f9237fa57671167dbcd1f9c45b8