tast-test: Add ShillInitLogoutScript test
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockFlimflam instead of shill to run
the logout test. MockFlimflam provides a subset of shill DBus API
so that we can track "dbus-send" invocations sent by the shill init
scripts.  These tests are moved to run on real shill in the new Tast
test, because it's possible to observe the expected behaviors without
the mock and that allows for checking the status of the profile stack as
part of these tests.

BUG=chromium:1026901
TEST=run tast network.ShillInitLogoutScript

Change-Id: I31f9031d3c021f465c647bc290f8f6c2e0379cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2050727
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>