tast-test: Add ShillInitScriptsLoginMultiProfile
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockFlimflam instead of shill to run
the login multi profile test. MockFlimflam provides a subset of shill DBus API
so that we can track "dbus-send" invocations sent by the shill init
scripts.  These tests are moved to run on real shill in the new Tast
test, because it's possible to observe the expected behaviors without
the mock and that allows for checking the status of the profile stack as
part of these tests.

BUG=chromium:1026901
TEST=run tast network.ShillInitScriptsLoginMultiProfile

Change-Id: I80b847c036b526972c72d66c543b123e35c06ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2051527
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>