tast-test: Add ShillInitScriptsLoginProfileExists
The Tast test is migrated from network_ShillInitScripts.py, an Autotest.
In the original version, it uses MockFlimflam instead of shill to run
the login profile exists test. MockFlimflam provides a subset of shill DBus API
so that we can track "dbus-send" invocations sent by the shill init
scripts.  These tests are moved to run on real shill in the new Tast
test, because it's possible to observe the expected behaviors without
the mock and that allows for checking the status of the profile stack as
part of these tests.

BUG=chromium:1026901
TEST=run tast network.ShillInitScriptsLoginProfileExists

Change-Id: I5ae125235dc695d9d832abf0973d87ac8342c2f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2050355
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>