network: UdevRename: stop "udevadm monitor" with SIGTERM
We read and wait on stdout of "udevadm monitor" in background. However,
as the process may buffer its stdout (especially on the models with few
devices/events), we should kill it with SIGTERM instead of SIGKILL on
timeout so that the process can at least flush its buffer.

BUG=chromium:1031943
TEST=`tast run $DUT network.UdevRename` on veyron_fievel

Change-Id: I140b7b8786ad1e7090f2cf176b22d32fd3a191bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2058942
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>