tast: Add new test for ARC printing
The new test adds a virtual USB printer and installs ArcPrintTest to
launch ARC printing. It then interacts with Chrome print preview to
select the virtual USB printer, change print settings, and print a
document. The printed document is compared to a golden document to
verify the content was printed correctly.

BUG=b:146647611
TEST=tast run --build $DUT_IP arc.Print

Change-Id: I95e348c458e8950621e2df45bdd2522565c34025