Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> LGTM (+1 since I'm the shadow reviewer).
> 
> I'm not sure how this var is for; my understanding is that the var can only be used for your local use (except for the secret variables), are you running those tests with custom variables locally?
> 
> Maybe I am missing something, but I am feeling like the use of parameterized tests would be great for  further analyses of changing the number of restarts (cf. https://chromium.googlesource.com/chromiumos/platform/tast-tests.git/+/refs/heads/master/src/chromiumos/tast/local/bundles/cros/ui/chrome_login.go#26).
> 
> Sorry if this comment is pointless.

This is indeed for testing locally.

Basically, I want to run this for ~1000 restarts to manually stress induce and seek out issues (and make it easy for others to do so in the future).  That might take ~2-3h and not worth doing for every build.  Hopefully once this is stable and reliable, we'd investigate any failures closely and make up for the low number of restarts with the large number of boards.

Patch-set: 3
Reviewer: Gerrit User 1108834 <1108834@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>