network: IwlwifiPCIRescan: move network.IwlwifiPCIRescan to remote
When the test failed, WiFi interface will be removed and this affects
later tests. We have a basic recovery mechanism (reload the driver) but
it doesn't seem to work very well. In the worst case, we might have to
reboot DUT to have the interface back. With current Tast design, we
cannot do this in a local test. Move this test to remote and add
recovery as suggested in [1].

Have a feature request in crbug.com/1051801. We might move this back to
local test if we can do similar recovery there.

[1] https://groups.google.com/a/chromium.org/g/tast-users/c/9ohr0Z9C8Nw

BUG=chromium:1051352
TEST=`tast run $DUT network.IwlwifiPCIRescan` on cyan DUT.
Also manually modify the rescan script to noop and verify the reboot
recovery does work.

Change-Id: Ide4c06cae5d1db1a973c64e15d9155a29679f4ec