platform: storage: Update SATA failure check for wearout test
Currently the test uses attribute fail indicators to determine
if a SATA device is failing. The attributes reported vary for devices,
so we should check the Percentage Used Endurance Indicator.

Also, some eMMC devices have a ']' in the DEVICE_LIFE_TIME_EST string,
so fix the regex to account for that.

BUG=chromium:1052162
TEST=Ran platform.StorageWearoutDetect on SATA, eMMC, and NVMe devices
Created example storage_info.txt files to check different behaviors

Change-Id: I1958c81260da360a63a6a613bf9c821ec2c3b122