Update patch set 3
Patch Set 3:

For the cryptohome readiness, crrev/c/2006568 does a very similar thing. I believe, via DaemonController we should not worry the code duplication, now and in the future. Sometimes we do have to restart cryptohome daemon (e.g., for soft-clear TPM or check verify if daemon is working after restart).

BTW, may I understand why we delete those OOBE-related files in tast? If they are not getting cleared along with clearing TPM ownership, is it a bug or WAI? If it is a bug we should fix it, and if it is WAI maybe we should have another test item that uses a proper way to power-wash the device and verify if those flag files are actually wiped?

Patch-set: 3