tast-tests: arc: Add arc.Reboot and arc.Reboot.vm tests
This is a port of cheets_ContainerReboot autotest. It checks whether
Android can be successfully rebooted. Reboot command is executed via
ADB. The success of reboot is checked by seeing whether ARC init PID is
changed after reboot. Note that in this test Android reboots only once,
whereas in the original autotest it reboots multiple times.

BUG=b:142574527
TEST=~/trunk/src/platform/tast/fast_build.sh -T
TEST=tast run <DUT-IP-eve> arc.Reboot/arc.Reboot.vm

Change-Id: If5a9786a5d7ff1eb2b929898fc7ced9e64dcb06a